+++ aeneas/syncmap.py
@@ -36,7 +36,7 @@ __copyright__ = """
     Copyright 2015,      Alberto Pettarin (www.albertopettarin.it)
     """
 __license__ = "GNU AGPL v3"
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 __email__ = "aeneas@readbeyond.it"
 __status__ = "Production"
 
@@ -55,6 +55,35 @@ class SyncMap(object):
     objects.
     """
 
+    FINETUNEAS_REPLACEMENTS = [
+        ["<!-- AENEAS_REPLACE_COMMENT_BEGIN -->", "<!-- AENEAS_REPLACE_COMMENT_BEGIN"],
+        ["<!-- AENEAS_REPLACE_COMMENT_END -->", "AENEAS_REPLACE_COMMENT_END -->"],
+        ["<!-- AENEAS_REPLACE_UNCOMMENT_BEGIN", "<!-- AENEAS_REPLACE_UNCOMMENT_BEGIN -->"],
+        ["AENEAS_REPLACE_UNCOMMENT_END -->", "<!-- AENEAS_REPLACE_UNCOMMENT_END -->"],
+        ["// AENEAS_REPLACE_SHOW_ID", "showID = true;"],
+        ["// AENEAS_REPLACE_ALIGN_TEXT", "alignText = \"left\""],
+        ["// AENEAS_REPLACE_CONTINUOUS_PLAY", "continuousPlay = true;"],
+        ["// AENEAS_REPLACE_TIME_FORMAT", "timeFormatHHMMSSmmm = true;"],
+    ]
+    FINETUNEAS_REPLACE_AUDIOFILEPATH = "// AENEAS_REPLACE_AUDIOFILEPATH"
+    FINETUNEAS_REPLACE_FRAGMENTS = "// AENEAS_REPLACE_FRAGMENTS"
+    FINETUNEAS_REPLACE_OUTPUT_FORMAT = "// AENEAS_REPLACE_OUTPUT_FORMAT"
+    FINETUNEAS_REPLACE_SMIL_AUDIOREF = "// AENEAS_REPLACE_SMIL_AUDIOREF"
+    FINETUNEAS_REPLACE_SMIL_PAGEREF = "// AENEAS_REPLACE_SMIL_PAGEREF"
+    FINETUNEAS_ALLOWED_FORMATS = [
+        "csv",
+        "json",
+        "smil",
+        "srt",
+        "ssv",
+        "ttml",
+        "tsv",
+        "txt",
+        "vtt",
+        "xml"
+    ]
+    FINETUNEAS_PATH = "res/finetuneas.html"
+
     TAG = "SyncMap"
 
     def __init__(self, logger=None):
@@ -72,6 +101,27 @@ class SyncMap(object):
 
     def __str__(self):
         return "\n".join([str(f) for f in self.fragments])
+
+    def to_json(self):
+        """
+        Return a JSON representation of the sync map.
+
+        :rtype: str
+
+        .. versionadded:: 1.3.1
+        """
+        output_fragments = []
+        for fragment in self.fragments:
+            text = fragment.text_fragment
+            output_fragment = {}
+            output_fragment["id"] = text.identifier
+            output_fragment["language"] = text.language
+            output_fragment["lines"] = text.lines
+            output_fragment["begin"] = gf.time_to_ssmmm(fragment.begin)
+            output_fragment["end"] = gf.time_to_ssmmm(fragment.end)
+            output_fragments.append(output_fragment)
+        output_dict = {"fragments": output_fragments}
+        return json.dumps(output_dict, indent=1, sort_keys=True)
 
     def append_fragment(self, fragment):
         """
@@ -104,6 +154,70 @@ class SyncMap(object):
         """
         self._log("Clearing sync map")
         self.fragments = []
+
+    def output_html_for_tuning(
+            self,
+            audio_file_path,
+            output_file_path,
+            parameters=None
+    ):
+        """
+        Output an HTML file for fine tuning the sync map manually.
+
+        :param audio_file_path: the path to the associated audio file
+        :type  audio_file_path: string (path)
+        :param output_file_path: the path to the output file to write
+        :type  output_file_path: string (path)
+        :param parameters: additional parameters
+        :type  parameters: dict
+
+        .. versionadded:: 1.3.1
+        """
+        if not gf.file_can_be_written(output_file_path):
+            raise IOError("Cannot output HTML file '%s' (wrong permissions?)" % output_file_path)
+        if parameters is None:
+            parameters = {}
+        audio_file_path_absolute = os.path.abspath(audio_file_path)
+        if os.name != "posix":
+            # on non-POSIX OSes, change e.g. c:\abc\def into c:/abc/def
+            # the browser will figure it out correctly
+            # TODO is there a better way to do this?
+            audio_file_path_absolute = audio_file_path_absolute.replace("\\", "/")
+        template_path_absolute = gf.get_rel_path(
+            self.FINETUNEAS_PATH,
+            from_path=os.path.dirname(__file__),
+            absolute=True
+        )
+        template = codecs.open(template_path_absolute, "r", "utf-8").read()
+        for repl in self.FINETUNEAS_REPLACEMENTS:
+            template = template.replace(repl[0], repl[1])
+        template = template.replace(
+            self.FINETUNEAS_REPLACE_AUDIOFILEPATH,
+            "audioFilePath = \"file://%s\";" % audio_file_path_absolute
+        )
+        template = template.replace(
+            self.FINETUNEAS_REPLACE_FRAGMENTS,
+            "fragments = (%s).fragments;" % self.to_json()
+        )
+        if gc.PPN_TASK_OS_FILE_FORMAT in parameters:
+            output_format = parameters[gc.PPN_TASK_OS_FILE_FORMAT]
+            if output_format in self.FINETUNEAS_ALLOWED_FORMATS:
+                template = template.replace(
+                    self.FINETUNEAS_REPLACE_OUTPUT_FORMAT,
+                    "outputFormat = \"%s\";" % output_format
+                )
+                if output_format == "smil":
+                    if gc.PPN_TASK_OS_FILE_SMIL_AUDIO_REF in parameters:
+                        template = template.replace(
+                            self.FINETUNEAS_REPLACE_SMIL_AUDIOREF,
+                            "audioref = \"%s\";" % parameters[gc.PPN_TASK_OS_FILE_SMIL_AUDIO_REF]
+                        )
+                    if gc.PPN_TASK_OS_FILE_SMIL_PAGE_REF in parameters:
+                        template = template.replace(
+                            self.FINETUNEAS_REPLACE_SMIL_PAGEREF,
+                            "pageref = \"%s\";" % parameters[gc.PPN_TASK_OS_FILE_SMIL_PAGE_REF]
+                        )
+        codecs.open(output_file_path, "w", "utf-8").write(template)
 
     def read(self, sync_map_format, input_file_path, parameters=None):
         """
@@ -352,18 +466,7 @@ class SyncMap(object):
         """
         Write to JSON file
         """
-        output_fragments = []
-        for fragment in self.fragments:
-            text = fragment.text_fragment
-            output_fragment = {}
-            output_fragment["id"] = text.identifier
-            output_fragment["language"] = text.language
-            output_fragment["lines"] = text.lines
-            output_fragment["begin"] = gf.time_to_ssmmm(fragment.begin)
-            output_fragment["end"] = gf.time_to_ssmmm(fragment.end)
-            output_fragments.append(output_fragment)
-        output_dict = {"fragments": output_fragments}
-        output_file.write(json.dumps(output_dict, indent=1, sort_keys=True))
+        output_file.write(self.to_json())
 
     def _read_rbse(self, input_file):
         """