+++ aeneas/tools/download.py
@@ -20,7 +20,7 @@ __copyright__ = """
     Copyright 2015,      Alberto Pettarin (www.albertopettarin.it)
     """
 __license__ = "GNU AGPL 3"
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 __email__ = "aeneas@readbeyond.it"
 __status__ = "Production"
 
@@ -58,11 +58,11 @@ def usage():
 
 def main():
     """ Entry point """
-
     if ("-h" in sys.argv) or ("--help" in sys.argv):
         # show full help
         usage()
-
+    if len(sys.argv) < 3:
+        usage()
     verbose = False
     best_audio = True
     source_url = sys.argv[1]